@@ -3,70 +3,63 @@
   "name": "propel-web",
   "version": "2.1.0",
   "description": "Propel Frontend",
-  "main": "main.js",
-  "dependencies": {
-    "@angular/animations": "^13.2.4",
-    "@angular/cdk": "^13.2.4",
-    "@angular/common": "^13.2.4",
-    "@angular/compiler": "^13.2.4",
-    "@angular/core": "^13.2.4",
-    "@angular/forms": "^13.2.4",
-    "@angular/material": "^13.2.4",
-    "@angular/platform-browser": "^13.2.4",
-    "@angular/platform-browser-dynamic": "^13.2.4",
-    "@angular/router": "^13.2.4",
-    "@ng-select/ng-select": "^8.1.1",
-    "@swimlane/ngx-charts": "^19.0.1",
-    "bootstrap": "^4.6.0",
-    "component": "^1.1.0",
-    "jquery": "^3.6.0",
-    "ngx-page-scroll": "^8.0.0",
-    "ngx-page-scroll-core": "^8.0.0",
-    "ngx-toastr": "^12.1.0",
-    "prismjs": "^1.25.0",
-    "rxjs": "~6.5.4",
-    "stemmer": "^1.0.5",
-    "stopword": "^1.0.11",
-    "tslib": "^2.3.1",
-    "zone.js": "~0.11.4"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "^13.2.5",
-    "@angular/cli": "^13.2.5",
-    "@angular/compiler-cli": "^13.2.4",
-    "@fortawesome/fontawesome-free": "^5.15.4",
-    "@types/jasmine": "^3.6.11",
-    "@types/jasminewd2": "^2.0.10",
-    "@types/node": "^12.20.26",
-    "@types/prismjs": "^1.16.6",
-    "@types/shortid": "0.0.29",
-    "@types/stemmer": "^1.0.2",
-    "@types/stopword": "^0.3.0",
-    "codelyzer": "^6.0.0",
-    "electron": "^10.4.7",
-    "electron-builder": "^22.11.7",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.4",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.7.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "^6.1.3",
-    "typescript": "^4.5.5"
-  },
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "startElectron": "electron .",
     "build": "ng build",
+    "watch": "ng build --watch --configuration development",
     "test": "ng test --watch=true",
-    "lint": "ng lint",
-    "e2e": "ng e2e",
     "electronBuild": "electron-builder build"
   },
-  "author": "JC",
-  "license": "ISC"
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^14.2.0",
+    "@angular/cdk": "^14.2.5",
+    "@angular/common": "^14.2.0",
+    "@angular/compiler": "^14.2.0",
+    "@angular/core": "^14.2.0",
+    "@angular/forms": "^14.2.0",
+    "@angular/material": "^14.2.5",
+    "@angular/platform-browser": "^14.2.0",
+    "@angular/platform-browser-dynamic": "^14.2.0",
+    "@angular/router": "^14.2.0",
+    "@ng-select/ng-select": "^9.0.2",
+    "@swimlane/ngx-charts": "^19.0.1",
+    "bootstrap": "^4.6.0",
+    "jquery": "^3.6.0",
+    "ngx-page-scroll": "^9.0.0",
+    "ngx-page-scroll-core": "^9.0.0",
+    "ngx-toastr": "^15.2.0",
+    "prismjs": "^1.29.0",
+    "rxjs": "~7.5.0",
+    "stemmer": "^1.0.5",
+    "stopword": "^1.0.11",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.11.4"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "^14.2.6",
+    "@angular/cli": "~14.2.6",
+    "@angular/compiler-cli": "^14.2.0",
+    "@fortawesome/fontawesome-free": "^5.15.4",
+    "@types/jasmine": "~4.0.0",
+    "@types/jasminewd2": "^2.0.10",
+    "@types/node": "^14.14.31",
+    "@types/prismjs": "^1.16.6",
+    "@types/shortid": "0.0.29",
+    "@types/stemmer": "^1.0.2",
+    "@types/stopword": "^0.3.0",
+    "electron": "^10.4.7",
+    "electron-builder": "^22.11.7",
+    "jasmine-core": "~4.3.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.0.0",
+    "ts-node": "^8.3.0",
+    "tslint": "^6.1.3",
+    "typescript": "~4.7.2"
+  }
 }
