@@ -28,62 +28,62 @@ import { NgxPageScrollModule } from 'ngx-page-scroll';
 import { NgxChartsModule } from '@swimlane/ngx-charts';
 
 //Services & Others
-import { CoreService } from './services/core.service';
-import { NavigationService } from './services/navigation.service';
-import { AppPages } from './services/app-pages.service';
-import { ErrorHandlerService } from "./services/error-handler.service";
-import { ToasterService } from './services/toaster.service';
-import { DataService } from './services/data.service';
-import { DialogService } from "./services/dialog.service";
-import { RunnerService } from './services/runner.service';
-import { LoaderInterceptor } from './core/loader-interceptor';
-import { AuthInterceptor } from './core/auth-interceptor';
-import { DataLossPreventionGuard } from './core/data-loss-prevention-guard';
-import { ParamInferenceService } from "./services/param-inference.service";
-import { InfiniteScrollingModule } from './core/infinite-scrolling-module';
-import { APIStatusService } from "./services/api-status.service";
+import { CoreService } from '../services/core.service';
+import { NavigationService } from '../services/navigation.service';
+import { AppPages } from '../services/app-pages.service';
+import { ErrorHandlerService } from "../services/error-handler.service";
+import { ToasterService } from '../services/toaster.service';
+import { DataService } from '../services/data.service';
+import { DialogService } from "../services/dialog.service";
+import { RunnerService } from '../services/runner.service';
+import { LoaderInterceptor } from '../core/loader-interceptor';
+import { AuthInterceptor } from '../core/auth-interceptor';
+import { DataLossPreventionGuard } from '../core/data-loss-prevention-guard';
+import { ParamInferenceService } from "../services/param-inference.service";
+import { InfiniteScrollingModule } from '../core/infinite-scrolling-module';
+import { APIStatusService } from "../services/api-status.service";
 
 //Dialogs
-import { WorkflowStepDialogComponent } from './app/dialogs/workflow-step-dlg/workflow-step-dlg.component';
-import { CustomFieldDialogComponent } from './app/dialogs/custom-field-dlg/custom-field-dlg.component';
-import { CustomValueDialogComponent } from './app/dialogs/custom-value-dlg/custom-value-dlg.component';
+import { WorkflowStepDialogComponent } from './dialogs/workflow-step-dlg/workflow-step-dlg.component';
+import { CustomFieldDialogComponent } from './dialogs/custom-field-dlg/custom-field-dlg.component';
+import { CustomValueDialogComponent } from './dialogs/custom-value-dlg/custom-value-dlg.component';
 
 //Components
-import { NavigationBarComponent } from './app/navigation-bar/navigation-bar.component';
-import { HomeComponent } from './app/home/home.component';
-import { SandboxComponent } from './app/sandbox/sandbox.component';
-import { RootComponent } from './app/root/root.component';
-import { RunComponent } from './app/run/run.component';
-import { TargetComponent } from './app/target/target.component';
-import { ValidationMessageComponent } from './app/validation-message/validation-message.component';
-import { ScriptComponent } from './app/script/script.component';
-import { FileDropperComponent } from './app/file-dropper/file-dropper.component';
-import { ScriptParametersComponent } from './app/script-parameters/script-parameters.component';
-import { QuickTaskComponent } from './app/quick-task/quick-task.component';
-import { WorkflowStepComponent } from './app/workflow-step/workflow-step.component';
-import { WorkflowComponent } from './app/workflow/workflow.component';
-import { ConsoleLineComponent } from './app/console-line/console-line.component';
-import { StatusIconComponent } from './app/status-icon/status-icon.component';
-import { ResultsComponent } from './app/results/results.component';
-import { DynamicTableComponent } from './app/dynamic-table/dynamic-table.component';
-import { SearchComponent } from './app/search/search.component';
-import { SearchWorkflowLineComponent } from './app/search-workflow-line/search-workflow-line.component';
-import { SearchScriptLineComponent } from './app/search-script-line/search-script-line.component';
-import { SearchTargetLineComponent } from './app/search-target-line/search-target-line.component';
-import { HistoryComponent } from './app/history/history.component';
-import { OfflineComponent } from './app/offline/offline.component';
-import { UnauthorizedComponent } from './app/unauthorized/unauthorized.component';
-import { CredentialComponent } from './app/credential/credential.component';
-import { WindowsSecretComponent } from './app/credential/windows-secret/windows-secret.component';
-import { AWSSecretComponent } from "./app/credential/aws-secret/aws-secret.component";
-import { SearchCredentialLineComponent } from './app/search-credential-line/search-credential-line.component';
-import { APIKeySecretComponent } from './app/credential/apikey-secret/apikey-secret.component';
-import { UserAccountComponent } from './app/user-account/user-account.component';
-import { SearchUserAccountLineComponent } from './app/search-user-account-line/search-user-account-line.component';
-import { SecurityService } from './services/security.service';
-import { AuthenticationCodeComponent } from './app/authentication-code/authentication-code.component';
-import { LoginComponent } from './app/login/login.component';
-import { SecurityGuard } from './services/security-guard';
+import { NavigationBarComponent } from './navigation-bar/navigation-bar.component';
+import { HomeComponent } from './home/home.component';
+import { SandboxComponent } from './sandbox/sandbox.component';
+import { RootComponent } from './root/root.component';
+import { RunComponent } from './run/run.component';
+import { TargetComponent } from './target/target.component';
+import { ValidationMessageComponent } from './validation-message/validation-message.component';
+import { ScriptComponent } from './script/script.component';
+import { FileDropperComponent } from './file-dropper/file-dropper.component';
+import { ScriptParametersComponent } from './script-parameters/script-parameters.component';
+import { QuickTaskComponent } from './quick-task/quick-task.component';
+import { WorkflowStepComponent } from './workflow-step/workflow-step.component';
+import { WorkflowComponent } from './workflow/workflow.component';
+import { ConsoleLineComponent } from './console-line/console-line.component';
+import { StatusIconComponent } from './status-icon/status-icon.component';
+import { ResultsComponent } from './results/results.component';
+import { DynamicTableComponent } from './dynamic-table/dynamic-table.component';
+import { SearchComponent } from './search/search.component';
+import { SearchWorkflowLineComponent } from './search-workflow-line/search-workflow-line.component';
+import { SearchScriptLineComponent } from './search-script-line/search-script-line.component';
+import { SearchTargetLineComponent } from './search-target-line/search-target-line.component';
+import { HistoryComponent } from './history/history.component';
+import { OfflineComponent } from './offline/offline.component';
+import { UnauthorizedComponent } from './unauthorized/unauthorized.component';
+import { CredentialComponent } from './credential/credential.component';
+import { WindowsSecretComponent } from './credential/windows-secret/windows-secret.component';
+import { AWSSecretComponent } from "./credential/aws-secret/aws-secret.component";
+import { SearchCredentialLineComponent } from './search-credential-line/search-credential-line.component';
+import { APIKeySecretComponent } from './credential/apikey-secret/apikey-secret.component';
+import { UserAccountComponent } from './user-account/user-account.component';
+import { SearchUserAccountLineComponent } from './search-user-account-line/search-user-account-line.component';
+import { SecurityService } from '../services/security.service';
+import { AuthenticationCodeComponent } from './authentication-code/authentication-code.component';
+import { LoginComponent } from './login/login.component';
+import { SecurityGuard } from '../services/security-guard';
 
 @NgModule({
     declarations: [
